# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for ShipStation application",
        "title": "ShipStation RestAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "http://localhost:8080",
    "basePath": "/",
    "paths": {
        "/request": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список всех не удаленных заявок",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявка"
                ],
                "summary": "Получение списка заявок",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Начало даты формирования",
                        "name": "startFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Конец даты формирования",
                        "name": "endFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Статус заявки",
                        "name": "requestStatus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заявок",
                        "schema": {
                            "$ref": "#/definitions/model.GetRequests"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/request/{requestID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает информацию о заявке по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявка"
                ],
                "summary": "Получение заявки по идентификатору",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор заявки",
                        "name": "requestID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о заявке",
                        "schema": {
                            "$ref": "#/definitions/model.GetRequestByID"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет заявку по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявка"
                ],
                "summary": "Удаление заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор заявки",
                        "name": "requestID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Номер рейса",
                        "name": "searchFlightNumber",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Начало даты формирования",
                        "name": "startFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Конец даты формирования",
                        "name": "endFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Статус заявки",
                        "name": "requestStatus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заявок",
                        "schema": {
                            "$ref": "#/definitions/model.GetRequests"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/request/{requestID}/status/moderator": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет статус заявки для модератора по идентификатору заявки",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявка"
                ],
                "summary": "Обновление статуса заявки для модератора",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор заявки",
                        "name": "requestID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус заявки",
                        "name": "requestStatus",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.StatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о заявке",
                        "schema": {
                            "$ref": "#/definitions/model.GetRequestByID"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/request/{requestID}/status/user": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет статус заявки для пользователя по идентификатору заявки",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Заявка"
                ],
                "summary": "Обновление статуса заявки для пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор заявки",
                        "name": "requestID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о заявке",
                        "schema": {
                            "$ref": "#/definitions/model.GetRequestByID"
                        }
                    },
                    "400": {
                        "description": "Обработанная ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "model.GetRequestByID": {
            "type": "object",
            "properties": {
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "request_id": {
                    "type": "integer"
                },
                "request_status": {
                    "type": "string"
                },
                "ships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Ship"
                    }
                }
            }
        },
        "model.GetRequests": {
            "type": "object",
            "properties": {
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "request_id": {
                    "type": "integer"
                },
                "request_status": {
                    "type": "string"
                }
            }
        },
...
        "model.Ship": {
            "type": "object",
            "properties": {
                "cargo_capacity": {
                    "type": "number"
                },
                "classification": {
                    "type": "string"
                },
                "crew_capacity": {
                    "type": "integer"
                },
                "flag": {
                    "type": "string"
                },
                "max_depth": {
                    "type": "number"
                },
                "max_length": {
                    "type": "number"
                },
                "passenger_capacity": {
                    "type": "integer"
                },
                "photo": {
                    "type": "string"
                },
                "ship_id": {
                    "type": "integer"
                },
                "ship_name": {
                    "type": "string"
                },
                "ship_status": {
                    "type": "string"
                },
                "ship_type": {
                    "type": "string"
                },
                "year_built": {
                    "type": "integer"
                }
            }
        },
...
        "model.StatusRequest": {
            "type": "object",
            "properties": {
                "request_status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}